+++ app/src/main/java/app/musikus/usecase/recordings/Types.kt
@@ -25,5 +25,5 @@ data class Recording(
 interface RecordingsRepository {
     val recordings: Flow<List<Recording>>
 
-    suspend fun getRawRecording(contentUri: Uri): Result<FloatArray>
+    suspend fun getRawRecording(contentUri: Uri): Result<ShortArray>
 }
